tests/wpt/web-platform-tests/css/css-color-adjust/inheritance.html

<!DOCTYPE html>
<title>Inheritance of CSS Color Adjust properties</title>
<link rel="help" href="https://drafts.csswg.org/css-color-adjust/#property-index">
<meta name="assert" content="Properties inherit according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
  assert_inherited('color-scheme', 'normal', 'light dark');
</script>
